requirements.txt
flask==2.3.3
flask-pymongo==2.3.0
flask-login==0.6.2
pymongo==4.5.0
python-dotenv==1.0.0
pandas==2.0.3
werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.0
bson==0.5.10
